Change part numbers based upon workshop
The workshop states "part 3 allocates memory both on the host and device", however in the code this is down as step 4. Same thing for "Part 4 communicates directly with the device and consists of copying initial values of vectors A and B to device memory", this has been changed from Step 5 to Step 4.